2D Node and 2D LinkList class
The LL2DNode class represents a node in the 2D linked list structure. Each LL2DNode contains a pointer to an instance of the LL class, representing the linked list of moves for a specific game session. These nodes are used to create a linked list of linked lists.

Public Functions (in LL2DNode Class):

LL2DNode() - Constructor for LL2DNode class.
Purpose: Initializes the LL2DNode object with default values, setting the next, prev, and current pointers to nullptr.
LL2D Class Description:
The LL2D class represents the 2D linked list that stores the entire session's moves. It contains a linked list of LL2DNode objects, each representing a game session's moves as a linked list.

Public Functions (in LL2D Class):

LL2D() - Constructor for LL2D class.

Purpose: Initializes the LL2D object with a nullptr head, indicating an empty 2D linked list.
addToList(LL* newList) - Adds a new linked list (representing a game session's moves) to the 2D linked list.

Parameters:
newList: A pointer to an instance of the LL class representing the moves for a game session.
Purpose: Adds a new game session's moves to the 2D linked list as an LL2DNode.
getHead() - Getter function to retrieve the head of the 2D linked list.

Purpose: Returns a pointer to the head node of the 2D linked list.
Class Usage:
The LL2D class is used to store and manage the history of game sessions in the Tic Tac Toe Pro game. Each node in the 2D linked list (represented by LL2DNode) corresponds to a specific game session and contains an instance of the LL class, which stores the moves for that session. This 2D structure allows for easy navigation through different game sessions and their moves.

It provides an organized way to keep track of multiple game sessions and facilitates viewing and managing the moves made during each session. The addToList function is used to add a new game session's moves to the 2D linked list, ensuring that the entire gaming history can be accessed and displayed when needed.